update: refatorando rota de question
